Create Day - 10 (calculator project)
In Day 10 I have done the project for calculation where anyone can use as a calculator tool. This Python script defines a simple calculator that supports addition, subtraction, multiplication, and division. The calculator operates in a command-line interface and allows users to perform multiple operations in sequence. The main function, calculator(), initializes the calculator, displays available operations, takes user input, performs calculations, and allows the user to continue or exit.