Create Day-11 (Black jack game)
In Day 11 I created a game called Blackjack this game is more familiar  who plays cards often. Blackjack game code implements a simple command-line version of the game using Python. It includes functions for dealing cards, calculating scores, and comparing results between the player and the computer. The game loop allows the player to draw cards until they choose to stop or bust, then the computer plays automatically according to standard rules, and the game outcomes are compared and displayed.