Create Day 15 (Coffee machine simulator)
In Day 15 Coffee Machine Simulator, This Python program simulates a coffee machine that can make espresso, latte, or cappuccino. It checks if sufficient resources are available, processes user payments, and updates the machine’s profit and resources accordingly. Users can request a coffee, view a report of current resources and earnings, or turn off the machine.